requirements.txt
python-telegram-bot==21.6
APScheduler==3.10.4
aiosqlite==0.20.0
python-dotenv==1.0.1